requirements.txt
pip~=21.3.1
attrs~=22.1.0
wheel~=0.37.1
Pillow~=9.2.0
tornado~=6.2
Jinja2~=3.1.2
filelock~=3.13.4
setuptools~=60.2.0
packaging~=21.3
zipp~=3.9.0
platformdirs~=4.2.1
flask~=3.0.3
werkzeug~=3.0.3
stripe~=9.6.0
python-dotenv~=1.0.1
flask-mail==0.9.1
Flask-SQLAlchemy==3.0.2
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
flask-principal==0.4.0
flask-login==0.6.3
flask-swagger-ui==4.11.1
pytest==8.2.0
gunicorn==22.0.0